feat: support preferredDatePickerStyle in iOS 14 (#246)
BREAKING CHANGE: you need XCode 11 to build this version, and you need this version to properly support iOS 14 (for that you need XCode 12)

This adds support for preferredDatePickerStyle to the module. We took the android-only display prop and extended the support for iOS too, as opposed to adding a new ios-only prop.