Fix incorrect size computation
Fix issue where UIDatePicker sizeThatFits: doesn't account for the bottom 44 height of the time picker row when configured with datePickerMode == UIDatePickerModeDateAndTime and datePickerStyle == UIDatePickerStyleInline. There are a few examples online of others running into this where wrapping in a parent view and configuring autolayout to the parent anchors fixes the issue but that did not seem like a proper fix for how UIDatePicker is used here. I've isolated this change ONLY to this use case as other calls to get size appear correct.