fix(android): improve color resource handling in Common.java (#913)
The changes handle the fetching of color resources more accurately in our Common.java file. The old method was simplified to one line, but this didn't consider raw color values. Now, the code correctly handles both cases: when the type value is resolved to a color resource ID, and when it's resolved to a raw color value.

Co-authored-by: Solera <cvalverde@irvinecompany.com>