feat(android): material 3 pickers (#952)
* Support Kotlin files

Android Studio recommended these settings for supporting
Kotlin in the project.

* Add Material dependency

* Extract re-usable picker argument utilities

The new Material pickers will utilize these utilities as well.

* Set up new specs/modules

* Accept new arguments for Material pickers

The pickers will allow the user to set a title for the dialog
and specify if the initial input mode should be a text input or
calendar/clock.

* Implement MaterialDatePicker

This also adds a utility function to RNDate to get the date in
milliseconds.

* Implement MaterialTimePicker

* Create JS connectors that call the native modules

* Choose native module based on `design` prop

If the `design` is "material", then use the Material modules.
Otherwise, use the default modules.

* Validate Material 3 props not used with default pickers

* Update components to pass new props to modules

* Update types

* Update example app

* Fix typo in existing docs

* Add dummy files to prevent tests from failing

The tests were trying to access the native modules. These dummy
files will make sure the tests don't try accessing the real
implementations that end with `.android.js`.

* Update README

* Update E2E specs

Since there are multiple "default" options on the screen now,
I've specified to tap the first one. I also had to make sure to
scroll the screen before tapping the show picker button.

* Make example app inherit from Material3 theme

* Move post-install into `start` script